requirements.txt
Django==3.2.2
django-import-export==2.5.0
mysqlclient==2.0.3
